@@ -1,14 +1,12 @@
 {
   "name": "@mdn/browser-compat-data",
-  "version": "3.2.4",
+  "version": "4.2.1",
   "description": "Browser compatibility data provided by MDN Web Docs",
   "main": "index.js",
   "types": "index.d.ts",
-  "dependencies": {
-    "extend": "3.0.2"
-  },
+  "type": "module",
   "engines": {
-    "node": ">=10.0.0"
+    "node": ">=16.0.0"
   },
   "repository": {
     "type": "git",
@@ -30,37 +28,55 @@
   },
   "homepage": "https://github.com/mdn/browser-compat-data#readme",
   "devDependencies": {
-    "ajv": "~6.12.2",
-    "better-ajv-errors": "~0.7.0",
-    "chalk": "~4.1.0",
-    "compare-versions": "~3.6.0",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.10.1",
-    "eslint-config-standard": "^14.1.1",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jsdoc": "^22.1.0",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-prefer-arrow": "^1.1.7",
-    "eslint-plugin-prettier": "^3.1.2",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.1",
+    "@babel/eslint-parser": "~7.17.0",
+    "@babel/plugin-syntax-import-assertions": "~7.17.12",
+    "@desertnet/html-parser": "~1.0.1",
+    "ajv": "~8.11.0",
+    "ajv-errors": "~3.0.0",
+    "ajv-formats": "~2.1.1",
+    "better-ajv-errors": "~1.2.0",
+    "browser-specs": "~3.11.0",
+    "c8": "~7.11.2",
+    "chalk": "~5.0.1",
+    "chalk-template": "~0.4.0",
+    "compare-versions": "~4.1.1",
+    "deep-diff": "~1.0.2",
+    "es-main": "~1.0.2",
+    "eslint": "~8.15.0",
+    "eslint-config-prettier": "~8.5.0",
+    "eslint-config-standard": "~17.0.0",
+    "eslint-plugin-import": "~2.26.0",
+    "eslint-plugin-jsdoc": "~39.3.0",
+    "eslint-plugin-node": "~11.1.0",
+    "eslint-plugin-prefer-arrow": "~1.2.3",
+    "eslint-plugin-prettier": "~4.0.0",
+    "eslint-plugin-promise": "~6.0.0",
+    "fast-json-stable-stringify": "~2.1.0",
+    "fdir": "~5.2.0",
     "mdn-confluence": "~2.2.2",
-    "mocha": "~8.3.0",
-    "ora": "~5.4.0",
-    "prettier": "~2.2.0",
-    "yargs": "~16.2.0"
+    "mocha": "~10.0.0",
+    "open-cli": "~7.0.1",
+    "ora": "~6.1.0",
+    "prettier": "~2.6.1",
+    "yargs": "~17.5.0"
   },
   "scripts": {
     "confluence": "node ./node_modules/mdn-confluence/main/generate.es6.js --output-dir=. --bcd-module=./index.js",
-    "mocha": "mocha \"*/**.test.js\"",
+    "diff": "node scripts/diff.js",
+    "unittest": "c8 mocha --recursive \"{,!(node_modules)/**}/*.test.js\"",
+    "coverage": "c8 report -r lcov && open-cli coverage/lcov-report/index.html",
+    "format": "npx eslint --fix \"**/*.js\" && npx prettier --check \"**/*.ts\" \"**/*.md\" \"**/*.json\"",
+    "format:fix": "npx eslint --fix \"**/*.js\" && npx prettier --write \"**/*.ts\" \"**/*.md\" \"**/*.json\"",
     "lint": "node test/lint",
-    "fix": "node scripts/fix",
+    "fix": "npm run format:fix && node scripts/fix",
     "mirror": "node scripts/mirror",
     "stats": "node scripts/statistics",
-    "release-pulls": "node scripts/release-pulls",
-    "release-stats": "node scripts/release-stats",
+    "release-build": "node scripts/release/build",
+    "release-pulls": "node scripts/release/pulls",
+    "release-notes": "node scripts/release/notes",
+    "release-stats": "node scripts/release/stats",
     "show-errors": "npm test 1> /dev/null",
-    "test": "npm run mocha && npm run lint",
+    "test": "npm run format && npm run lint && npm run unittest",
     "traverse": "node scripts/traverse"
   }
 }
