@@ -1,194 +1,183 @@
-'use strict';
-const fs = require('fs');
-const path = require('path');
-const ora = require('ora');
-const yargs = require('yargs');
-const chalk = require('chalk');
-const {
-  testBrowsers,
-  testLinks,
-  testPrefix,
-  testRealValues,
-  testStyle,
-  testSchema,
-  testVersions,
+/* This file is a part of @mdn/browser-compat-data
+ * See LICENSE file for more information. */
+
+import fs from 'node:fs';
+import path from 'node:path';
+import { fileURLToPath } from 'node:url';
+
+import esMain from 'es-main';
+import ora from 'ora';
+import yargs from 'yargs';
+import { hideBin } from 'yargs/helpers';
+import chalk from 'chalk-template';
+
+import {
+  testBrowsersData,
+  testBrowsersPresence,
   testConsistency,
   testDescriptions,
-} = require('./linter/index.js');
-const { IS_CI } = require('./utils.js');
-const testCompareFeatures = require('./test-compare-features');
-const testMigrations = require('./test-migrations');
-const testFormat = require('./test-format');
+  testLinks,
+  testNotes,
+  testPrefix,
+  testSchema,
+  testStyle,
+  testVersions,
+} from './linter/index.js';
+import { IS_CI, pluralize } from './utils.js';
 
-/** @type {Map<string, string>} */
-const filesWithErrors = new Map();
+const dirname = fileURLToPath(new URL('.', import.meta.url));
 
-const argv = yargs
-  .alias('version', 'v')
-  .usage('$0 [[--] files...]', false, yargs => {
-    return yargs.positional('files...', {
-      description: 'The files to lint',
-      type: 'string',
-    });
-  })
-  .help()
-  .alias('help', 'h')
-  .alias('help', '?')
-  .parse(process.argv.slice(2));
+/** @type {object} */
+const spinner = ora({
+  stream: process.stdout,
+});
 
 /**
- * @param {string[]} files
- * @return {boolean}
+ * Recursively checks files for any errors.
+ *
+ * @param {string[]} files The files to test
+ * @returns {object} Errors by relative file path.
  */
-function load(...files) {
-  return files.reduce((prevHasErrors, file) => {
-    if (file.indexOf(__dirname) !== 0) {
-      file = path.resolve(__dirname, '..', file);
+const checkFiles = (...files) => {
+  let errors = {};
+  for (let file of files) {
+    if (file.indexOf(dirname) !== 0) {
+      file = path.resolve(dirname, '..', file);
     }
 
     if (!fs.existsSync(file)) {
-      return prevHasErrors; // Ignore non-existent files
+      console.warn(chalk`{yellow File {bold ${file}} doesn't exist!}`);
+      return;
     }
 
-    if (fs.statSync(file).isFile()) {
-      let fileHasErrors = false;
+    if (fs.statSync(file).isFile() && path.extname(file) === '.json') {
+      const filePath = {
+        full: path.relative(process.cwd(), file),
+      };
+      filePath.category =
+        filePath.full.includes(path.sep) && filePath.full.split(path.sep)[0];
 
-      if (path.extname(file) === '.json') {
-        let hasSyntaxErrors = false,
-          hasSchemaErrors = false,
-          hasStyleErrors = false,
-          hasLinkErrors = false,
-          hasBrowserErrors = false,
-          hasVersionErrors = false,
-          hasConsistencyErrors = false,
-          hasRealValueErrors = false,
-          hasPrefixErrors = false,
-          hasDescriptionsErrors = false;
-        const relativeFilePath = path.relative(process.cwd(), file);
+      spinner.text = filePath.full;
 
-        const spinner = ora({
-          stream: process.stdout,
-          text: relativeFilePath,
-        });
+      if (!IS_CI) {
+        // Continuous integration environments don't allow overwriting
+        // previous lines using VT escape sequences, which is how
+        // the spinner animation is implemented.
+        spinner.start();
+      }
 
-        if (!IS_CI) {
-          // Continuous integration environments don't allow overwriting
-          // previous lines using VT escape sequences, which is how
-          // the spinner animation is implemented.
-          spinner.start();
-        }
-
-        const console_error = console.error;
-        console.error = (...args) => {
+      // Catch console errors and report them as file errors
+      const console_error = console.error;
+      console.error = (...args) => {
+        if (!(filePath.full in errors)) {
+          // Set spinner to failure when first error is found
+          // Setting on every error causes duplicate output
           spinner['stream'] = process.stderr;
-          spinner.fail(chalk.red.bold(relativeFilePath));
-          console.error = console_error;
-          console.error(...args);
-        };
+          spinner.fail(chalk`{red.bold ${filePath.full}}`);
 
-        try {
-          if (file.indexOf('browsers' + path.sep) !== -1) {
-            hasSchemaErrors = testSchema(
-              file,
-              './../../schemas/browsers.schema.json',
-            );
-            hasLinkErrors = testLinks(file);
-          } else {
-            hasSchemaErrors = testSchema(file);
-            hasStyleErrors = testStyle(file);
-            hasLinkErrors = testLinks(file);
-            hasBrowserErrors = testBrowsers(file);
-            hasVersionErrors = testVersions(file);
-            hasConsistencyErrors = testConsistency(file);
-            hasRealValueErrors = testRealValues(file);
-            hasPrefixErrors = testPrefix(file);
-            hasDescriptionsErrors = testDescriptions(file);
-          }
-        } catch (e) {
-          hasSyntaxErrors = true;
-          console.error(e);
+          errors[filePath.full] = [];
         }
+        console_error(...args);
+        errors[filePath.full].push(...args);
+      };
 
-        fileHasErrors = [
-          hasSyntaxErrors,
-          hasSchemaErrors,
-          hasStyleErrors,
-          hasLinkErrors,
-          hasBrowserErrors,
-          hasVersionErrors,
-          hasConsistencyErrors,
-          hasRealValueErrors,
-          hasPrefixErrors,
-          hasDescriptionsErrors,
-        ].some(x => !!x);
+      try {
+        const rawFileData = fs
+          .readFileSync(new URL(file, import.meta.url), 'utf-8')
+          .trim();
+        const fileData = JSON.parse(rawFileData);
 
-        if (fileHasErrors) {
-          filesWithErrors.set(relativeFilePath, file);
+        testSchema(fileData, filePath);
+        testLinks(rawFileData);
+
+        if (file.indexOf('browsers' + path.sep) !== -1) {
+          testBrowsersData(fileData);
         } else {
-          console.error = console_error;
-          spinner.succeed();
+          testBrowsersPresence(fileData, filePath);
+          testConsistency(fileData);
+          testDescriptions(fileData);
+          testPrefix(fileData, filePath);
+          testStyle(rawFileData);
+          testVersions(fileData);
+          testNotes(fileData);
         }
+      } catch (e) {
+        console.error(e);
       }
 
-      return prevHasErrors || fileHasErrors;
+      // Reset console.error
+      console.error = console_error;
+
+      if (!(filePath.full in errors)) {
+        spinner.succeed();
+      }
     }
 
-    const subFiles = fs.readdirSync(file).map(subfile => {
-      return path.join(file, subfile);
-    });
+    if (fs.statSync(file).isDirectory()) {
+      const subFiles = fs.readdirSync(file).map((subfile) => {
+        return path.join(file, subfile);
+      });
 
-    return load(...subFiles) || prevHasErrors;
-  }, false);
-}
-
-/** @type {boolean} */
-var hasErrors = argv.files
-  ? load.apply(undefined, argv.files)
-  : load(
-      'api',
-      'browsers',
-      'css',
-      'html',
-      'http',
-      'svg',
-      'javascript',
-      'mathml',
-      'webdriver',
-      'webextensions',
-      'xpath',
-      'xslt',
-    );
-hasErrors = testCompareFeatures() || hasErrors;
-hasErrors = testMigrations() || hasErrors;
-hasErrors = testFormat() || hasErrors;
-
-if (hasErrors) {
-  console.warn('');
-  console.warn(
-    chalk`{red Problems in {bold ${filesWithErrors.size}} ${
-      filesWithErrors.size === 1 ? 'file' : 'files'
-    }:}`,
-  );
-  for (const [fileName, file] of filesWithErrors) {
-    console.warn(chalk`{red.bold ✖ ${fileName}}`);
-    try {
-      if (file.indexOf('browsers' + path.sep) !== -1) {
-        testSchema(file, './../../schemas/browsers.schema.json');
-        testLinks(file);
-      } else {
-        testSchema(file);
-        testStyle(file);
-        testLinks(file);
-        testVersions(file);
-        testRealValues(file);
-        testBrowsers(file);
-        testConsistency(file);
-        testPrefix(file);
-        testDescriptions(file);
-      }
-    } catch (e) {
-      console.error(e);
+      errors = { ...errors, ...checkFiles(...subFiles) };
     }
   }
-  process.exit(1);
+
+  return errors;
+};
+
+/**
+ * Test for any errors in specified file(s) and/or folder(s), or all of BCD
+ *
+ * @param {?string} files The file(s) and/or folder(s) to test. Leave null for everything.
+ * @returns {boolean} Whether there were any errors
+ */
+const main = (
+  files = [
+    'api',
+    'browsers',
+    'css',
+    'html',
+    'http',
+    'svg',
+    'javascript',
+    'mathml',
+    'webdriver',
+    'webextensions',
+  ],
+) => {
+  const errors = checkFiles(...files);
+
+  const filesWithErrors = Object.keys(errors).length;
+
+  if (filesWithErrors) {
+    console.error('');
+    console.error(
+      chalk`{red Problems in {bold ${pluralize('file', filesWithErrors)}}:}`,
+    );
+
+    for (const [fp, errorMsgs] of Object.entries(errors)) {
+      console.error(chalk`{red.bold ✖ ${fp}}`);
+      for (const error of errorMsgs) {
+        console.error(error);
+      }
+    }
+  }
+
+  return filesWithErrors > 0;
+};
+
+if (esMain(import.meta)) {
+  const { argv } = yargs(hideBin(process.argv)).command(
+    '$0 [files..]',
+    false,
+    (yargs) =>
+      yargs.positional('files...', {
+        description: 'The files to lint (leave blank to test everything)',
+        type: 'string',
+      }),
+  );
+
+  process.exit(main(argv.files) ? 1 : 0);
 }
+
+export default main;
