@@ -1,7 +1,10 @@
-'use strict';
-const path = require('path');
-const compareVersions = require('compare-versions');
-const chalk = require('chalk');
+/* This file is a part of @mdn/browser-compat-data
+ * See LICENSE file for more information. */
+
+import compareVersions from 'compare-versions';
+import chalk from 'chalk-template';
+import { query } from '../../utils/index.js';
+import { Logger } from '../utils.js';
 
 /**
  * @typedef {import('../../types').CompatStatement} CompatStatement
@@ -15,15 +18,15 @@ const chalk = require('chalk');
  * @typedef {'unsupported' | 'support_unknown' | 'subfeature_earlier_implementation'} ErrorType
  *
  * @typedef {object} ConsistencyError
- * @property {string} feature
- * @property {string[]} path
- * @property {FeatureError[]} errors
+ * @property {string} feature The identifier of the feature
+ * @property {string[]} path The path of the feature
+ * @property {FeatureError[]} errors Any errors found
  *
  * @typedef {object} FeatureError
- * @property {ErrorType} errortype
- * @property {string} browser
- * @property {VersionValue} parent_value
- * @property {[string, VersionValue][]} subfeatures
+ * @property {ErrorType} errortype The type of the error
+ * @property {string} browser The browser the error was found
+ * @property {VersionValue} parent_value The value of the parent feature
+ * @property {[string, VersionValue][]} subfeatures The versions of the subfeatures
  */
 
 /**
@@ -32,19 +35,23 @@ const chalk = require('chalk');
  * This checker aims at improving data quality
  * by detecting inconsistent information.
  */
-class ConsistencyChecker {
+export class ConsistencyChecker {
   /**
-   * @param {Identifier} data
-   * @return {ConsistencyError[]}
+   * Checks the data for any errors
+   *
+   * @param {Identifier} data The data to test
+   * @returns {ConsistencyError[]} Any errors found within the data
    */
   check(data) {
     return this.checkSubfeatures(data);
   }
 
   /**
-   * @param {Identifier} data
-   * @param {string[]} [path]
-   * @return {ConsistencyError[]}
+   * Recursively checks the data for any errors
+   *
+   * @param {Identifier} data The data to test
+   * @param {string[]} [path] The path of the data
+   * @returns {ConsistencyError[]} Any errors found within the data
    */
   checkSubfeatures(data, path = []) {
     /** @type {ConsistencyError[]} */
@@ -52,13 +59,10 @@ class ConsistencyChecker {
 
     // Check this feature.
     if (this.isFeature(data)) {
-      const feature = path.length ? path[path.length - 1] : 'ROOT';
-
       const errors = this.checkFeature(data);
 
       if (errors.length) {
         allErrors.push({
-          feature,
           path,
           errors,
         });
@@ -66,11 +70,13 @@ class ConsistencyChecker {
     }
 
     // Check sub-features.
-    const keys = Object.keys(data).filter(key => key != '__compat');
-    keys.forEach(key => {
+    this.getSubfeatures(data).forEach((key) => {
       allErrors = [
         ...allErrors,
-        ...this.checkSubfeatures(data[key], [...path, key]),
+        ...this.checkSubfeatures(query(key, data), [
+          ...path,
+          ...key.split('.'),
+        ]),
       ];
     });
 
@@ -79,15 +85,37 @@ class ConsistencyChecker {
 
   /**
    * @param {PrimaryIdentifier & Required<IdentifierMeta>} data
-   * @return {FeatureError[]}
+   * @return {Identifier[]}
+   */
+  getSubfeatures(data) {
+    const subfeatures = [];
+    const keys = Object.keys(data).filter((key) => key != '__compat');
+    for (const key of keys) {
+      if ('__compat' in data[key]) {
+        // If the subfeature has compat data
+        subfeatures.push(key);
+      } else {
+        // If no compat data, get the next level down
+        subfeatures.push(
+          ...this.getSubfeatures(data[key]).map((x) => `${key}.${x}`),
+        );
+      }
+    }
+
+    return subfeatures;
+  }
+
+  /**
+   * Checks a specific feature for errors
+   *
+   * @param {Identifier} data The data to test
+   * @returns {FeatureError[]} Any errors found within the data
    */
   checkFeature(data) {
     /** @type {FeatureError[]} */
     let errors = [];
 
-    const subfeatures = Object.keys(data).filter(key =>
-      this.isFeature(data[key]),
-    );
+    const subfeatures = this.getSubfeatures(data);
 
     // Test whether sub-features are supported when basic support is not implemented
     // For all unsupported browsers (basic support == false), sub-features should be set to false
@@ -95,20 +123,21 @@ class ConsistencyChecker {
     /** @type {Partial<Record<string, [string, VersionValue][]>>} */
     var inconsistentSubfeaturesByBrowser = {};
 
-    subfeatures.forEach(subfeature => {
+    subfeatures.forEach((subfeature) => {
       const unsupportedInChild = this.extractUnsupportedBrowsers(
-        data[subfeature].__compat,
+        query(subfeature, data).__compat,
       );
 
       const browsers = unsupportedInParent.filter(
-        x => !unsupportedInChild.includes(x),
+        (x) => !unsupportedInChild.includes(x),
       );
 
-      browsers.forEach(browser => {
+      browsers.forEach((browser) => {
         inconsistentSubfeaturesByBrowser[browser] =
           inconsistentSubfeaturesByBrowser[browser] || [];
-        const subfeature_value =
-          data[subfeature].__compat.support[browser].version_added;
+        const subfeature_value = this.getVersionAdded(
+          query(subfeature, data).__compat.support[browser],
+        );
         inconsistentSubfeaturesByBrowser[browser].push([
           subfeature,
           subfeature_value,
@@ -117,10 +146,10 @@ class ConsistencyChecker {
     });
 
     // Add errors
-    Object.keys(inconsistentSubfeaturesByBrowser).forEach(browser => {
+    Object.keys(inconsistentSubfeaturesByBrowser).forEach((browser) => {
       const subfeatures = inconsistentSubfeaturesByBrowser[browser];
       const errortype = 'unsupported';
-      const parent_value = data.__compat.support[browser].version_added;
+      const parent_value = this.getVersionAdded(data.__compat.support[browser]);
 
       errors.push({
         errortype,
@@ -137,20 +166,21 @@ class ConsistencyChecker {
     );
     inconsistentSubfeaturesByBrowser = {};
 
-    subfeatures.forEach(subfeature => {
+    subfeatures.forEach((subfeature) => {
       const supportUnknownInChild = this.extractSupportNotTrueBrowsers(
-        data[subfeature].__compat,
+        query(subfeature, data).__compat,
       );
 
       const browsers = supportUnknownInParent.filter(
-        x => !supportUnknownInChild.includes(x),
+        (x) => !supportUnknownInChild.includes(x),
       );
 
-      browsers.forEach(browser => {
+      browsers.forEach((browser) => {
         inconsistentSubfeaturesByBrowser[browser] =
           inconsistentSubfeaturesByBrowser[browser] || [];
-        const subfeature_value =
-          data[subfeature].__compat.support[browser].version_added;
+        const subfeature_value = this.getVersionAdded(
+          query(subfeature, data).__compat.support[browser],
+        );
         inconsistentSubfeaturesByBrowser[browser].push([
           subfeature,
           subfeature_value,
@@ -159,10 +189,10 @@ class ConsistencyChecker {
     });
 
     // Add errors
-    Object.keys(inconsistentSubfeaturesByBrowser).forEach(browser => {
+    Object.keys(inconsistentSubfeaturesByBrowser).forEach((browser) => {
       const subfeatures = inconsistentSubfeaturesByBrowser[browser];
       const errortype = 'support_unknown';
-      const parent_value = data.__compat.support[browser].version_added;
+      const parent_value = this.getVersionAdded(data.__compat.support[browser]);
 
       errors.push({
         errortype,
@@ -178,19 +208,20 @@ class ConsistencyChecker {
     );
     inconsistentSubfeaturesByBrowser = {};
 
-    subfeatures.forEach(subfeature => {
-      supportInParent.forEach(browser => {
+    subfeatures.forEach((subfeature) => {
+      supportInParent.forEach((browser) => {
         if (
-          data[subfeature].__compat.support[browser] != undefined &&
+          query(subfeature, data).__compat.support[browser] != undefined &&
           this.isVersionAddedGreater(
-            data[subfeature].__compat.support[browser],
+            query(subfeature, data).__compat.support[browser],
             data.__compat.support[browser],
           )
         ) {
           inconsistentSubfeaturesByBrowser[browser] =
             inconsistentSubfeaturesByBrowser[browser] || [];
-          const subfeature_value =
-            data[subfeature].__compat.support[browser].version_added;
+          const subfeature_value = this.getVersionAdded(
+            query(subfeature, data).__compat.support[browser],
+          );
           inconsistentSubfeaturesByBrowser[browser].push([
             subfeature,
             subfeature_value,
@@ -200,10 +231,10 @@ class ConsistencyChecker {
     });
 
     // Add errors
-    Object.keys(inconsistentSubfeaturesByBrowser).forEach(browser => {
+    Object.keys(inconsistentSubfeaturesByBrowser).forEach((browser) => {
       const subfeatures = inconsistentSubfeaturesByBrowser[browser];
       const errortype = 'subfeature_earlier_implementation';
-      const parent_value = data.__compat.support[browser].version_added;
+      const parent_value = this.getVersionAdded(data.__compat.support[browser]);
 
       errors.push({
         errortype,
@@ -217,21 +248,25 @@ class ConsistencyChecker {
   }
 
   /**
-   * @param {Identifier} data
-   * @return {data is Required<IdentifierMeta>}
+   * Checks if the data is a feature
+   *
+   * @param {Identifier} data The data to test
+   * @returns {boolean} If the data is a feature statement
    */
   isFeature(data) {
     return '__compat' in data;
   }
 
   /**
-   * @param {CompatStatement} compatData
-   * @return {string[]}
+   * Get all of the unsupported browsers in a feature
+   *
+   * @param {CompatStatement} compatData The compat data to process
+   * @returns {string[]} The list of browsers marked as unsupported
    */
   extractUnsupportedBrowsers(compatData) {
     return this.extractBrowsers(
       compatData,
-      data =>
+      (data) =>
         data.version_added === false ||
         (typeof data.version_removed !== 'undefined' &&
           data.version_removed !== false),
@@ -239,24 +274,28 @@ class ConsistencyChecker {
   }
 
   /**
-   * @param {CompatStatement} compatData
-   * @return {string[]}
+   * Get all of the browsers with unknown support in a feature
+   *
+   * @param {CompatStatement} compatData The compat data to process
+   * @returns {string[]} The list of browsers with unknown support
    */
   extractSupportUnknownBrowsers(compatData) {
     return this.extractBrowsers(
       compatData,
-      data => data.version_added === null,
+      (data) => data.version_added === null,
     );
   }
 
   /**
-   * @param {CompatStatement} compatData
-   * @return {string[]}
+   * Get all of the browsers with either unknown or no support in a feature
+   *
+   * @param {CompatStatement} compatData The compat data to process
+   * @returns {string[]} The list of browsers with non-truthy (false or null) support
    */
   extractSupportNotTrueBrowsers(compatData) {
     return this.extractBrowsers(
       compatData,
-      data =>
+      (data) =>
         data.version_added === false ||
         data.version_added === null ||
         (typeof data.version_removed !== 'undefined' &&
@@ -264,52 +303,69 @@ class ConsistencyChecker {
     );
   }
   /**
-   * @param {CompatStatement} compatData
-   * @return {string[]}
+   * Get all of the browsers with a version number in a feature.
+   *
+   * @param {CompatStatement} compatData The compat data to process
+   * @returns {string[]} The list of browsers with an exact version number
    */
   extractSupportedBrowsersWithVersion(compatData) {
     return this.extractBrowsers(
       compatData,
-      data => typeof data.version_added === 'string',
+      (data) => typeof data.version_added === 'string',
     );
   }
 
   /**
-   * @param {SupportStatement} compatData
-   * @return {string | null}
+   * Return the earliest recorded version number from a support statement or null.
+   *
+   * @param {SupportStatement} supportstatement The compat data to process
+   * @returns {?string} The earliest version added in the data
    */
-  getVersionAdded(compatData) {
-    /** @type {string | null} */
-    var version_added = null;
+  getVersionAdded(supportStatement) {
+    // A convenience function to squash non-real values and previews into null
+    const resolveVersionAddedValue = (statement) =>
+      [true, false, 'preview', null].includes(statement?.version_added)
+        ? null
+        : statement?.version_added;
 
-    if (typeof compatData.version_added === 'string')
-      return compatData.version_added;
-
-    if (compatData.constructor === Array) {
-      for (var i = compatData.length - 1; i >= 0; i--) {
-        var va = compatData[i].version_added;
-        if (
-          typeof va === 'string' &&
-          (version_added == null ||
-            (typeof version_added === 'string' &&
-              compareVersions.compare(
-                version_added.replace('≤', ''),
-                va.replace('≤', ''),
-                '>',
-              )))
-        ) {
-          version_added = va;
-        }
-      }
+    // Handle simple support statements
+    if (!Array.isArray(supportStatement)) {
+      return resolveVersionAddedValue(supportStatement);
     }
 
-    return version_added;
+    // Handle array support statements
+    let selectedValue = null;
+    for (const statement of supportStatement) {
+      const resolvedValue = resolveVersionAddedValue(statement);
+
+      if (resolvedValue === null) {
+        // We're not going to get a more specific version, so bail out now
+        return null;
+      }
+
+      if (selectedValue !== null) {
+        // Earlier value takes precedence
+        const resolvedIsEarlier = compareVersions.compare(
+          resolvedValue.replace('≤', ''),
+          selectedValue.replace('≤', ''),
+          '<',
+        );
+        if (resolvedIsEarlier) {
+          selectedValue = resolvedValue;
+        }
+      } else {
+        selectedValue = resolvedValue;
+      }
+    }
+    return selectedValue;
   }
 
   /**
-   * @param {SupportStatement} a
-   * @param {SupportStatement} b
-   * @return {boolean}
+   * Compare two versions and determine if a's version is greater (later) than b's version
+   *
+   * @param {SupportStatement} a The first support statement to compare
+   * @param {SupportStatement} b The second support statement to compare
+   * @returns {boolean} If a's version is greater (later) than b's version
    */
   isVersionAddedGreater(a, b) {
     var a_version_added = this.getVersionAdded(a);
@@ -320,6 +376,15 @@ class ConsistencyChecker {
       typeof b_version_added === 'string'
     ) {
       if (b_version_added.startsWith('≤')) {
+        return false;
+      }
+      if (a_version_added === 'preview' && b_version_added === 'preview') {
+        return false;
+      }
+      if (b_version_added === 'preview') {
+        return true;
+      }
+      if (a_version_added === 'preview') {
         return false;
       }
       return compareVersions.compare(
@@ -333,13 +398,14 @@ class ConsistencyChecker {
   }
 
   /**
+   * Get all of the browsers within the data and pass the data to the callback.
    *
-   * @param {CompatStatement} compatData
-   * @param {(browserData: SimpleSupportStatement) => boolean} callback
-   * @return {string[]}
+   * @param {CompatStatement} compatData The compat data to process
+   * @param {(browserData: SimpleSupportStatement) => boolean} callback The function to pass the data to
+   * @returns {boolean} The result of the invoked callback, or "false"
    */
   extractBrowsers(compatData, callback) {
-    return Object.keys(compatData.support).filter(browser => {
+    return Object.keys(compatData.support).filter((browser) => {
       const browserData = compatData.support[browser];
 
       if (Array.isArray(browserData)) {
@@ -354,54 +420,38 @@ class ConsistencyChecker {
 }
 
 /**
- * @param {string} filename
+ * @param {Identifier} data The contents of the file to test
+ * @returns {boolean} If the file contains errors
  */
-function testConsistency(filename) {
-  /** @type {Identifier} */
-  let data = require(filename);
+export default function testConsistency(data) {
+  const logger = new Logger('Consistency');
 
   const checker = new ConsistencyChecker();
-  const errors = checker.check(data);
+  const allErrors = checker.check(data);
 
-  if (errors.length) {
-    console.error(
-      chalk`{red   Consistency - {bold ${errors.length} }${
-        errors.length === 1 ? 'error' : 'errors'
-      }:}`,
-    );
-    errors.forEach(({ feature, path, errors }) => {
-      console.error(
-        chalk`{red   → {bold ${errors.length}} × {bold ${feature}} [${path.join(
-          '.',
-        )}]: }`,
-      );
-      errors.forEach(({ errortype, browser, parent_value, subfeatures }) => {
-        if (errortype == 'unsupported') {
-          console.error(
-            chalk`{red     → No support in {bold ${browser}}, but support is declared in the following sub-feature(s):}`,
-          );
-        } else if (errortype == 'support_unknown') {
-          console.error(
-            chalk`{red     → Unknown support in parent for {bold ${browser}}, but support is declared in the following sub-feature(s):}`,
-          );
-        } else if (errortype == 'subfeature_earlier_implementation') {
-          console.error(
-            chalk`{red     → Basic support in {bold ${browser}} was declared implemented in a later version ({bold ${parent_value}}) than the following sub-feature(s):}`,
-          );
-        }
+  for (const { path, errors } of allErrors) {
+    let errorMessage = chalk`{bold ${errors.length}} × ${path
+      .slice(0, 1)
+      .join('.')}.{bold ${path[path.length - 1]}}: `;
+    for (const { errortype, browser, parent_value, subfeatures } of errors) {
+      if (errortype == 'unsupported') {
+        errorMessage += chalk`\n{red       → No support in {bold ${browser}}, but support is declared in the following sub-feature(s):}`;
+      } else if (errortype == 'support_unknown') {
+        errorMessage += chalk`\n{red       → Unknown support in parent for {bold ${browser}}, but support is declared in the following sub-feature(s):}`;
+      } else if (errortype == 'subfeature_earlier_implementation') {
+        errorMessage += chalk`\n{red       → Basic support in {bold ${browser}} was declared implemented in a later version ({bold ${parent_value}}) than the following sub-feature(s):}`;
+      }
 
-        subfeatures.forEach(subfeature => {
-          console.error(
-            chalk`{red       → {bold ${path.join('.')}.${subfeature[0]}}: ${
-              subfeature[1] === undefined ? '[Array]' : subfeature[1]
-            }}`,
-          );
-        });
-      });
-    });
-    return true;
+      for (const subfeature of subfeatures) {
+        errorMessage += chalk`\n{red         → {bold ${path.join('.')}.${
+          subfeature[0]
+        }}: ${subfeature[1] === undefined ? '[Array]' : subfeature[1]}}`;
+      }
+
+      logger.error(errorMessage);
+    }
   }
-  return false;
-}
 
-module.exports = testConsistency;
+  logger.emit();
+  return logger.hasErrors();
+}
