@@ -1,16 +1,28 @@
-'use strict';
-const fs = require('fs');
-const url = require('url');
-const chalk = require('chalk');
-const { IS_WINDOWS, indexToPos, indexToPosRaw } = require('../utils.js');
+/* This file is a part of @mdn/browser-compat-data
+ * See LICENSE file for more information. */
+
+import chalk from 'chalk-template';
+import { IS_WINDOWS, indexToPos, indexToPosRaw, Logger } from '../utils.js';
 
 /**
- * @param {string} filename
+ * @typedef {object} LinkError
+ * @property {string} issue The description of the error
+ * @property {[?number, ?number]} pos The cursor position of the issue in number-array form
+ * @property {string} posString The cursor position of the issue in string form
+ * @property {?string} expected The expected string if applicable
+ * @property {string} actualLink What the link currently is
  */
-function processData(filename) {
+
+/**
+ * Process the data for any errors within the links
+ *
+ * @param {string} rawData The raw contents of the file to test
+ * @returns {LinkError[]} A list of errors found in the links
+ */
+export function processData(rawData) {
   let errors = [];
 
-  let actual = fs.readFileSync(filename, 'utf-8').trim();
+  let actual = rawData;
 
   // prevent false positives from git.core.autocrlf on Windows
   if (IS_WINDOWS) {
@@ -22,7 +34,7 @@ function processData(filename) {
     errors,
     actual,
     String.raw`https?://bugzilla\.mozilla\.org/show_bug\.cgi\?id=(\d+)`,
-    match => {
+    (match) => {
       return {
         issue: 'Use shortenable URL',
         expected: `https://bugzil.la/${match[1]}`,
@@ -35,10 +47,23 @@ function processData(filename) {
     errors,
     actual,
     String.raw`https?://(bugs\.chromium\.org|code\.google\.com)/p/chromium/issues/detail\?id=(\d+)`,
-    match => {
+    (match) => {
       return {
         issue: 'Use shortenable URL',
         expected: `https://crbug.com/${match[2]}`,
+      };
+    },
+  );
+
+  processLink(
+    // use https://crbug.com/category/100000 instead
+    errors,
+    actual,
+    String.raw`https?://(bugs\.chromium\.org|code\.google\.com)/p/((?!chromium)\w+)/issues/detail\?id=(\d+)`,
+    (match) => {
+      return {
+        issue: 'Use shortenable URL',
+        expected: `https://crbug.com/${match[2]}/${match[3]}`,
       };
     },
   );
@@ -48,7 +73,7 @@ function processData(filename) {
     errors,
     actual,
     String.raw`https?://bugs\.webkit\.org/show_bug\.cgi\?id=(\d+)`,
-    match => {
+    (match) => {
       return {
         issue: 'Use shortenable URL',
         expected: `https://webkit.org/b/${match[1]}`,
@@ -61,7 +86,7 @@ function processData(filename) {
     errors,
     actual,
     String.raw`(....)<a href='((https?)://(bugzil\.la|crbug\.com|webkit\.org/b)/(\d+))'>(.*?)</a>`,
-    match => {
+    (match) => {
       const [, before, url, protocol, domain, bugId, linkText] = match;
 
       if (protocol !== 'https') {
@@ -102,7 +127,7 @@ function processData(filename) {
     errors,
     actual,
     String.raw`\b(https?)://((?:[a-z][a-z0-9-]*\.)*)developer.mozilla.org/(.*?)(?=["'\s])`,
-    match => {
+    (match) => {
       const [, protocol, subdomain, path] = match;
       const [, locale, expectedPath_] = /^(?:(\w\w(?:-\w\w)?)\/)?(.*)$/.exec(
         path,
@@ -111,7 +136,7 @@ function processData(filename) {
 
       if (!expectedPath.startsWith('docs/')) {
         // Convert legacy zone URLs (see https://bugzil.la/1462475):
-        const [zone, index] = /** @return {[string|null, number]} */ (() => {
+        const [zone, index] = /** @returns {[?string, number]} */ (() => {
           const match = expectedPath.match(
             /\b(Add-ons|Apps|Archive|Firefox|Learn|Web)\b/,
           );
@@ -159,7 +184,7 @@ function processData(filename) {
     errors,
     actual,
     String.raw`https?://developer.microsoft.com/(\w\w-\w\w)/(.*?)(?=["'\s])`,
-    match => {
+    (match) => {
       return {
         issue: 'Use non-localized Microsoft Developer URL',
         expected: `https://developer.microsoft.com/${match[2]}`,
@@ -171,8 +196,8 @@ function processData(filename) {
     errors,
     actual,
     String.raw`<a href='([^'>]+)'>((?:.(?!</a>))*.)</a>`,
-    match => {
-      if (url.parse(match[1]).hostname === null) {
+    (match) => {
+      if (new URL(match[1]).hostname === null) {
         return {
           issue: 'Include hostname in URL',
           actualLink: match[1],
@@ -186,10 +211,13 @@ function processData(filename) {
 }
 
 /**
- * @param {Object[]} errors
- * @param {string} actual
- * @param {string|RegExp} regexp
- * @param {(match: RegExpExecArray) => Object} matchHandler
+ * Given a RegEx expression, test the link for errors
+ *
+ * @param {LinkError[]} errors The errors object to push the new errors to
+ * @param {string} actual The link to test
+ * @param {string|RegExp} regexp The regex to test with
+ * @param {(match: Array.<?string>) => object} matchHandler The callback
+ * @returns {void}
  */
 function processLink(errors, actual, regexp, matchHandler) {
   const re = new RegExp(regexp, 'g');
@@ -214,26 +242,26 @@ function processLink(errors, actual, regexp, matchHandler) {
 }
 
 /**
- * @param {string} filename
+ * Test for any malformed links
+ *
+ * @param {string} rawData The raw contents of the file to test
+ * @returns {boolean} If the file contains errors
  */
-function testLinks(filename) {
+export default function testLinks(rawData) {
+  // XXX This should use the parsed JSON data
+
+  const logger = new Logger('Links');
+
   /** @type {Object[]} */
-  let errors = processData(filename);
+  let errors = processData(rawData);
 
-  if (errors.length) {
-    console.error(
-      chalk`{red   Links – {bold ${errors.length}} ${
-        errors.length === 1 ? 'error' : 'errors'
-      }:}`,
+  for (const error of errors) {
+    logger.error(
+      chalk`${error.posString} – ${error.issue} ({yellow ${error.actual}} → {green ${error.expected}}).`,
+      chalk`Run {bold npm run fix} to fix links automatically`,
     );
-    for (const error of errors) {
-      console.error(
-        chalk`  {red → ${error.posString} – ${error.issue} ({yellow ${error.actual}} → {green ${error.expected}}).}`,
-      );
-    }
-    return true;
   }
-  return false;
-}
 
-module.exports = testLinks;
+  logger.emit();
+  return logger.hasErrors();
+}
