@@ -1,48 +1,56 @@
-'use strict';
-const path = require('path');
-const chalk = require('chalk');
+/* This file is a part of @mdn/browser-compat-data
+ * See LICENSE file for more information. */
+
+import chalk from 'chalk-template';
+import { Logger } from '../utils.js';
 
 /**
  * @typedef {import('../../types').Identifier} Identifier
+ * @typedef {import('../utils').Logger} Logger
  */
 
 /**
- * @param {Identifier} data
- * @param {string} category
- * @param {string[]} errors
- * @param {string} prefix
- * @param {string} [path]
+ * Check the prefix of a specific feature
+ *
+ * @param {Identifier} data The data to test
+ * @param {string} category The category the data belongs to
+ * @param {string} prefix The browser-based prefix to test
+ * @param {Logger} logger The logger to output errors to
+ * @param {string} [path] The path to the data
+ * @returns {string[]} Any errors found within the data
  */
-function checkPrefix(data, category, errors, prefix, path = '') {
+function checkPrefix(data, category, prefix, logger, path = '') {
   for (const key in data) {
     if (key === 'prefix' && typeof data[key] === 'string') {
       if (data[key].includes(prefix)) {
-        const error = chalk`{red → {bold ${prefix}} prefix is wrong for key: {bold ${path}}}`;
         const rules = [
           category == 'api' && !data[key].startsWith(prefix),
           category == 'css' && !data[key].startsWith(`-${prefix}`),
         ];
-        if (rules.some(x => x === true)) {
-          errors.push(error);
+        if (rules.some((x) => x === true)) {
+          logger.error(
+            chalk`{bold ${prefix}} prefix is wrong for key: {bold ${path}}`,
+          );
         }
       }
     } else {
       if (typeof data[key] === 'object') {
         const curr_path = path.length > 0 ? `${path}.${key}` : key;
-        checkPrefix(data[key], category, errors, prefix, curr_path);
+        checkPrefix(data[key], category, prefix, logger, curr_path);
       }
     }
   }
-  return errors;
 }
 
 /**
- * @param {Identifier} data
- * @param {string} category
- * @return {string[]}
+ * Process the data for prefix errors
+ *
+ * @param {Identifier} data The data to test
+ * @param {string} category The category the data belongs to
+ * @param {Logger} logger The logger to output errors to
+ * @returns {void}
  */
-function processData(data, category) {
-  let errors = [];
+function processData(data, category, logger) {
   let prefixes = [];
 
   if (category === 'api') {
@@ -53,36 +61,22 @@ function processData(data, category) {
   }
 
   for (const prefix of prefixes) {
-    checkPrefix(data, category, errors, prefix);
+    checkPrefix(data, category, prefix, logger);
   }
-  return errors;
 }
 
 /**
- * @param {string} filename
+ * Test for issues with feature's prefix
+ *
+ * @param {Identifier} data The contents of the file to test
+ * @param {object} filePath The path info for the file being tested
+ * @returns {boolean} If the file contains errors
  */
-function testPrefix(filename) {
-  const relativePath = path.relative(
-    path.resolve(__dirname, '..', '..'),
-    filename,
-  );
-  const category =
-    relativePath.includes(path.sep) && relativePath.split(path.sep)[0];
-  const data = require(filename);
-  const errors = processData(data, category);
+export default function testPrefix(data, filePath) {
+  const logger = new Logger('Prefix');
 
-  if (errors.length) {
-    console.error(
-      chalk`{red   Prefix – {bold ${errors.length}} ${
-        errors.length === 1 ? 'error' : 'errors'
-      }:}`,
-    );
-    for (const error of errors) {
-      console.error(`  ${error}`);
-    }
-    return true;
-  }
-  return false;
+  processData(data, filePath.category, logger);
+
+  logger.emit();
+  return logger.hasErrors();
 }
-
-module.exports = testPrefix;
