@@ -1,6 +1,8 @@
-'use strict';
-const assert = require('assert');
-const { escapeInvisibles } = require('./utils.js');
+/* This file is a part of @mdn/browser-compat-data
+ * See LICENSE file for more information. */
+
+import assert from 'node:assert/strict';
+import { escapeInvisibles } from './utils.js';
 
 describe('utils', () => {
   it('`escapeInvisibles()` works correctly', () => {
@@ -49,7 +51,7 @@ describe('utils', () => {
       } else {
         [char, expected = char] = data;
       }
-      assert.strictEqual(escapeInvisibles(char), expected);
+      assert.equal(escapeInvisibles(char), expected);
     }
   });
 });
