@@ -1,4 +1,7 @@
-const bcd = require('..');
+/* This file is a part of @mdn/browser-compat-data
+ * See LICENSE file for more information. */
+
+import bcd from '../index.js';
 
 /**
  * Get a subtree of compat data.
@@ -8,7 +11,7 @@ const bcd = require('..');
  * @returns {*} A BCD subtree
  * @throws {ReferenceError} For invalid identifiers
  */
-function query(path, data = bcd) {
+export default function query(path, data = bcd) {
   const pathElements = path.split('.');
   let lookup = data;
   while (pathElements.length) {
@@ -22,5 +25,3 @@ function query(path, data = bcd) {
   }
   return lookup;
 }
-
-module.exports = query;
