@@ -1,12 +1,15 @@
-const assert = require('assert').strict;
+/* This file is a part of @mdn/browser-compat-data
+ * See LICENSE file for more information. */
 
-const visit = require('./visit');
-const { walk } = require('./walk');
+import assert from 'node:assert/strict';
+
+import visit from './visit.js';
+import walk from './walk.js';
 
 describe('visit()', function () {
   it('runs the function on all features if no other entry point is specified', function () {
     const walker = walk();
-    visit(visitorPath => {
+    visit((visitorPath) => {
       assert.equal(visitorPath, walker.next().value.path);
     });
   });
@@ -15,7 +18,7 @@ describe('visit()', function () {
     const hits = new Set();
     const misses = new Set();
     visit(
-      path => {
+      (path) => {
         hits.add(path);
       },
       {
@@ -36,7 +39,7 @@ describe('visit()', function () {
   });
 
   it('visitorFn can break iteration', function () {
-    visit(path => {
+    visit((path) => {
       if (path.startsWith('css')) {
         return visit.BREAK;
       }
@@ -49,7 +52,7 @@ describe('visit()', function () {
   });
 
   it('visitorFn can skip traversal of children', function () {
-    visit(path => {
+    visit((path) => {
       if (path === 'css.at-rules.counter-style') {
         return visit.CONTINUE;
       }
