@@ -1,8 +1,11 @@
-const assert = require('assert').strict;
+/* This file is a part of @mdn/browser-compat-data
+ * See LICENSE file for more information. */
 
-const bcd = require('..');
-const query = require('./query');
-const { joinPath, isBrowser, isFeature } = require('./walkingUtils');
+import assert from 'node:assert/strict';
+
+import bcd from '../index.js';
+import query from './query.js';
+import { joinPath, isBrowser, isFeature } from './walkingUtils.js';
 
 describe('joinPath()', function () {
   it('joins dotted paths to features', function () {
@@ -17,7 +20,7 @@ describe('joinPath()', function () {
 
 describe('isBrowser()', function () {
   it('returns true for browser-like objects', function () {
-    assert.ok(isBrowser(bcd.browsers.firefox));
+    assert.equal(isBrowser(bcd.browsers.firefox), true);
   });
 
   it('returns false for feature-like objects', function () {
@@ -31,6 +34,6 @@ describe('isFeature()', function () {
   });
 
   it('returns true for feature-like objects', function () {
-    assert.ok(isFeature(query('html.elements.a')));
+    assert.equal(isFeature(query('html.elements.a')), true);
   });
 });
