@@ -1,5 +1,5 @@
-/* Any copyright is dedicated to the Public Domain.
- * http://creativecommons.org/publicdomain/zero/1.0/ */
+/* This file is a part of @mdn/browser-compat-data
+ * See LICENSE file for more information. */
 
 /**
  * The names of the known browsers.
@@ -7,11 +7,13 @@
 export type BrowserNames =
   | 'chrome'
   | 'chrome_android'
+  | 'deno'
   | 'edge'
   | 'firefox'
   | 'firefox_android'
   | 'ie'
   | 'nodejs'
+  | 'oculus'
   | 'opera'
   | 'opera_android'
   | 'safari'
@@ -27,6 +29,8 @@ export type BrowserEngines =
   | 'Trident'
   | 'WebKit'
   | 'V8';
+
+export type BrowserTypes = 'desktop' | 'mobile' | 'xr' | 'server';
 
 /**
  * The browser namespace.
@@ -46,17 +50,39 @@ export interface BrowserStatement {
   name: string;
 
   /**
-   * The known versions of this browser.
+   * The platform the browser runs on, for example:
+   * `"desktop"`, `"mobile"`, `"server"`, etc.
    */
-  releases: {
-    [version: string]: ReleaseStatement;
-  };
+  type: BrowserTypes;
+
+  /**
+   * Whether the browser supports flags to enable or disable features.
+   */
+  accepts_flags?: boolean;
+
+  /**
+   * Whether the browser supports extensions.
+   */
+  accepts_webextensions?: boolean;
 
   /**
    * An optional string containing the URL of the page where feature flags can be changed
    * (e.g. `"about:config"` for Firefox or `"chrome://flags"` for Chrome).
    */
   pref_url?: string;
+
+  /**
+   * The preview browser's name, for example:
+   * `"Nightly"`, `"Canary"`, `"TP"`, etc.
+   */
+  preview_name: string;
+
+  /**
+   * The known versions of this browser.
+   */
+  releases: {
+    [version: string]: ReleaseStatement;
+  };
 }
 
 /**
@@ -375,16 +401,4 @@ interface CompatDataIdentifiers
    * JavaScript APIs and manifest keys.
    */
   webextensions: PrimaryIdentifier;
-
-  /**
-   * Contains data for [XPath](https://developer.mozilla.org/docs/Web/XPath)
-   * axes, and functions.
-   */
-  xpath: PrimaryIdentifier;
-
-  /**
-   * Contains data for [XSLT](https://developer.mozilla.org/docs/Web/XSLT)
-   * elements, attributes, and global attributes.
-   */
-  xslt: PrimaryIdentifier;
 }
