@@ -1,9 +1,12 @@
-const bcd = require('..');
-const { isBrowser, descendantKeys, joinPath } = require('./walkingUtils');
-const query = require('./query');
+/* This file is a part of @mdn/browser-compat-data
+ * See LICENSE file for more information. */
 
-function* lowLevelWalk(data = bcd, path, depth = Infinity) {
-  if (path !== undefined) {
+import bcd from '../index.js';
+import { isBrowser, descendantKeys, joinPath } from './walkingUtils.js';
+import query from './query.js';
+
+export function* lowLevelWalk(data = bcd, path, depth = Infinity) {
+  if (path !== undefined && path !== '__meta') {
     const next = {
       path,
       data,
@@ -24,7 +27,7 @@ function* lowLevelWalk(data = bcd, path, depth = Infinity) {
   }
 }
 
-function* walk(entryPoints, data = bcd) {
+export default function* walk(entryPoints, data = bcd) {
   const walkers = [];
 
   if (entryPoints === undefined) {
@@ -32,7 +35,7 @@ function* walk(entryPoints, data = bcd) {
   } else {
     entryPoints = Array.isArray(entryPoints) ? entryPoints : [entryPoints];
     walkers.push(
-      ...entryPoints.map(entryPoint =>
+      ...entryPoints.map((entryPoint) =>
         lowLevelWalk(query(entryPoint, data), entryPoint),
       ),
     );
@@ -46,5 +49,3 @@ function* walk(entryPoints, data = bcd) {
     }
   }
 }
-
-module.exports = { walk, lowLevelWalk };
