@@ -1,17 +1,26 @@
-function joinPath() {
+/* This file is a part of @mdn/browser-compat-data
+ * See LICENSE file for more information. */
+
+export function joinPath() {
   return Array.from(arguments).filter(Boolean).join('.');
 }
 
-function isFeature(obj) {
+export function isFeature(obj) {
   return '__compat' in obj;
 }
-function isBrowser(obj) {
+
+export function isBrowser(obj) {
   return 'name' in obj && 'releases' in obj;
 }
 
-function descendantKeys(data) {
+export function descendantKeys(data) {
+  if (typeof data !== 'object') {
+    // Return if the data isn't an object
+    return [];
+  }
+
   if (isFeature(data)) {
-    return Object.keys(data).filter(key => key !== '__compat');
+    return Object.keys(data).filter((key) => !key.startsWith('__'));
   }
 
   if (isBrowser(data)) {
@@ -19,12 +28,5 @@ function descendantKeys(data) {
     return [];
   }
 
-  return Object.keys(data);
+  return Object.keys(data).filter((key) => key !== '__meta');
 }
-
-module.exports = {
-  joinPath,
-  isFeature,
-  isBrowser,
-  descendantKeys,
-};
